webkit/data/layout_test_results/v8/LayoutTests/fast/js/function-toString-object-literals-expected.txt

This test checks that object literals are serialized properly. It's needed in part because JavaScriptCore converts numeric property names to string and back.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS compileAndSerialize('a = { 1: null }') is 'a = { 1: null }'
PASS compileAndSerialize('a = { 0: null }') is 'a = { 0: null }'
FAIL compileAndSerialize('a = { 1.0: null }') should be a = { 1: null }. Was a = { 1.0: null }.
PASS compileAndSerialize('a = { "1.0": null }') is 'a = { "1.0": null }'
FAIL compileAndSerialize('a = { 1e-500: null }') should be a = { 0: null }. Was a = { 1e-500: null }.
PASS compileAndSerialize('a = { 1e-300: null }') is 'a = { 1e-300: null }'
FAIL compileAndSerialize('a = { 1e300: null }') should be a = { 1e+300: null }. Was a = { 1e300: null }.
FAIL compileAndSerialize('a = { 1e500: null }') should be a = { "Infinity": null }. Was a = { 1e500: null }.
FAIL compileAndSerialize('a = { NaN: null }') should be a = { "NaN": null }. Was a = { NaN: null }.
FAIL compileAndSerialize('a = { Infinity: null }') should be a = { "Infinity": null }. Was a = { Infinity: null }.
FAIL compileAndSerialize('a = { "1": null }') should be a = { 1: null }. Was a = { "1": null }.
PASS compileAndSerialize('a = { "1hi": null }') is 'a = { "1hi": null }'
PASS compileAndSerialize('a = { "\'": null }') is 'a = { "\'": null }'
PASS compileAndSerialize('a = { "\\"": null }') is 'a = { "\\"": null }'
PASS compileAndSerialize('a = { get x() { } }') is 'a = { get x() { } }'
PASS compileAndSerialize('a = { set x(y) { } }') is 'a = { set x(y) { } }'
PASS compileAndSerialize('a = { --1: null }') threw exception SyntaxError: Unexpected token --.
PASS compileAndSerialize('a = { -NaN: null }') threw exception SyntaxError: Unexpected token -.
PASS compileAndSerialize('a = { -0: null }') threw exception SyntaxError: Unexpected token -.
PASS compileAndSerialize('a = { -0.0: null }') threw exception SyntaxError: Unexpected token -.
PASS compileAndSerialize('a = { -Infinity: null }') threw exception SyntaxError: Unexpected token -.
PASS successfullyParsed is true

TEST COMPLETE
